[DEV] process and model multiple scans of same session

+++ adr_dwi/workflows.py
@@ -19,6 +19,7 @@ import os
 import glob
 import shutil
 import json
+import toml
 import platform
 import pandas as pd
 import openpyxl  # noqa: F401
@@ -126,7 +127,12 @@ def clean_rawdata(data_dir: PT):
 
 
 def preproc_dwi(
-    subj: str, sess: str, data_dir: PT, work_dir: PT, log_dir: PT
+    subj: str,
+    sess: str,
+    data_dir: PT,
+    work_dir: PT,
+    log_dir: PT,
+    run: int = None,
 ) -> PT:
     """Conduct preprocessing of DWI via FSL.
 
@@ -147,6 +153,7 @@ def preproc_dwi(
         work_dir: Location for intermediates.
         data_dir: Location of BIDS organized directory.
         log_dir: Location for capturing STDOUT/ERR.
+        run: Optional, run ID for scan-rescan of same session.
 
     Returns:
         Location of eddy output file in data dir.
@@ -161,7 +168,11 @@ def preproc_dwi(
     log.write.info(f"Starting preproc_dwi: {subj}, {sess}")
 
     # Set output name and dir
-    out_name = f"{subj}_{sess}_dir-AP_desc-eddy_dwi.nii.gz"
+    out_name = (
+        f"{subj}_{sess}_dir-AP_desc-eddy_dwi.nii.gz"
+        if not run
+        else f"{subj}_{sess}_dir-AP_run-{run}_desc-eddy_dwi.nii.gz"
+    )
     out_dir = os.path.join(
         data_dir, "derivatives", "dwi_preproc", subj, sess, "dwi"
     )
@@ -173,8 +184,17 @@ def preproc_dwi(
     if os.path.exists(out_path):
         return out_path
 
+    # Set out names accounting for run use
+    comb_out = f"tmp_AP_PA_r{run}_b0" if run else "tmp_AP_PA_b0"
+    top_out = f"tmp_topup_r{run}" if run else "tmp_topup"
+    param_out = f"tmp_r{run}_acq_param.txt" if run else "tmp_acq_param.txt"
+    mask_out = f"tmp_r{run}_brain.nii.gz" if run else "tmp_brain.nii.gz"
+    idx_out = f"tmp_r{run}_index.txt" if run else "tmp_index.txt"
+
     # Setup for preprocessing
     dwi_pp = process.DwiPreproc(subj, sess, work_dir, data_dir)
+    if run:
+        dwi_pp.run = run
     dwi_dict, fmap_dict = dwi_pp.setup()
 
     # Get AP, PA files
@@ -183,22 +203,25 @@ def preproc_dwi(
         if "fmap" not in fmap_type:
             continue
         dir_val = fmap_type.split("_")[1]
-        b0_dict[f"b0_{dir_val}"] = dwi_pp.extract_b0(
-            fmap_path, f"tmp_{dir_val}_b0"
-        )
+        tmp_out = f"tmp_{dir_val}_r{run}_b0" if run else f"tmp_{dir_val}_b0"
+        b0_dict[f"b0_{dir_val}"] = dwi_pp.extract_b0(fmap_path, tmp_out)
 
     # Preprocess for topup
     ap_pa_b0 = dwi_pp.combine_b0(
-        b0_dict["b0_AP"], b0_dict["b0_PA"], "tmp_AP_PA_b0"
+        b0_dict["b0_AP"], b0_dict["b0_PA"], out_name=comb_out
     )
-    acq_param = dwi_pp.acq_param(fmap_dict["json_AP"])
+    acq_param = dwi_pp.acq_param(fmap_dict["json_AP"], out_name=param_out)
     dwi_topup, dwi_unwarp = dwi_pp.run_topup(
-        ap_pa_b0, acq_param, f"topup_{subj[4:]}_{sess[4:]}", log_dir
+        ap_pa_b0,
+        acq_param,
+        f"topup_{subj[4:]}_{sess[4:]}",
+        log_dir,
+        out_name=top_out,
     )
 
     # Preprocess with eddy
-    dwi_mask = dwi_pp.brain_mask(dwi_unwarp)
-    dwi_idx = dwi_pp.write_index(dwi_dict["dwi"])
+    dwi_mask = dwi_pp.brain_mask(dwi_unwarp, out_name=mask_out)
+    dwi_idx = dwi_pp.write_index(dwi_dict["dwi"], out_name=idx_out)
     dwi_eddy = dwi_pp.run_eddy(
         dwi_dict["dwi"],
         dwi_dict["bvec"],
@@ -224,6 +247,8 @@ def preproc_dwi(
     # Clean session dir
     if not os.path.exists(out_path):
         raise FileNotFoundError(out_path)
+    if run:
+        return
     shutil.rmtree(os.path.dirname(os.path.dirname(dwi_eddy)))
     log.write.info(f"Finished preproc_dwi: {subj}, {sess}")
     return out_path
@@ -269,6 +294,7 @@ def setup_pyafq(
     sess: str,
     data_dir: PT,
     work_dir: PT,
+    run: str = None,
 ) -> PT:
     """Setup for running pyAFQ.
 
@@ -283,6 +309,7 @@ def setup_pyafq(
         sess: BIDS session ID.
         data_dir: Location of BIDS organized directory.
         work_dir: Location for intermediates.
+        run: Optional, BIDS run ID for scan-rescan of same session.
 
     Raises:
         EnvironmentError: FSL not executable in system OS.
@@ -304,6 +331,8 @@ def setup_pyafq(
     out_mask = os.path.join(
         subj_work, f"{subj}_{sess}_dir-AP_desc-brain_mask.nii.gz"
     )
+    if run:
+        out_mask = out_mask.replace("_desc-brain", f"_{run}_desc-brain")
     if os.path.exists(out_mask):
         return out_mask
 
@@ -318,6 +347,9 @@ def setup_pyafq(
         "bval": f"{subj}_{sess}_dir-AP_dwi.bval",
         "json": f"{subj}_{sess}_dir-AP_dwi.json",
     }
+    if run:
+        for key, value in preproc_dict.items():
+            preproc_dict[key] = value.replace("_dir-AP", f"_dir-AP_{run}")
 
     # Pull files
     for file_type, file_name in preproc_dict.items():
@@ -357,6 +389,7 @@ def wrap_setup_pyafq(
     subj_sess: list,
     data_dir: PT,
     work_dir: PT,
+    run_list: list = None,
 ):
     """Submit setup_pyafq for array task ID.
 
@@ -368,6 +401,7 @@ def wrap_setup_pyafq(
         subj_sess: Tuples of BIDS subject, session IDs.
         data_dir: BIDS data location.
         work_dir: Location for intermediates.
+        run_list: Optional, list of run IDs for scan-rescan of same session.
 
     Raises:
         EnvironmentError: OS global variable 'SLURM_ARRAY_TASK_ID' not found.
@@ -393,15 +427,25 @@ def wrap_setup_pyafq(
     with open(os.path.join(work_deriv, "dataset_description.json"), "w") as jf:
         json.dump(data_desc, jf)
 
-    # Write config.toml
+    # Write config.toml from resources, update bids path
     log.write.info("Writing config.toml")
-    with pkg_resources.open_text(adr_bin, "config.toml") as cnf_in:
-        cnf_lines = cnf_in.read()
-    with open(os.path.join(work_deriv, "config.toml"), "w") as cnf_out:
-        cnf_out.write(cnf_lines)
+    config_data = toml.load(
+        pkg_resources.files(adr_bin).joinpath("config.toml").open("r")
+    )
+    config_data["BIDS_PARAMS"]["bids_path"] = work_deriv
+    work_config = os.path.join(work_deriv, "config.toml")
+    with open(work_config, "w") as cnf_out:
+        toml.dump(config_data, cnf_out)
 
     # Identify iteration subject and session list
     subj, sess = subj_sess[int(arr_id)]
+    if run_list:
+        for run_id in run_list:
+            run = f"run-{run_id}"
+            log.write.info(f"Starting setup_afq for: {subj}, {sess}, {run}")
+            _ = setup_pyafq(subj, sess, data_dir, work_dir, run=run)
+        return
+
     log.write.info(f"Starting setup_afq for: {subj}, {sess}")
     _ = setup_pyafq(subj, sess, data_dir, work_dir)
 
@@ -487,11 +531,12 @@ def run_pyafq(data_dir: PT, work_dir: PT, log_dir: PT, rerun: bool) -> PT:
     return out_path
 
 
-def insert_pyafq(rerun: bool = False) -> pd.DataFrame:
+def insert_pyafq(rerun: bool = False, rescan: bool = False) -> pd.DataFrame:
     """Insert data from pyAFQ tract_profiles.csv into db_adr.tbl_afq.
 
     Args:
-        rerun: Optional, get and send rerun metrics instead of all.
+        rerun: Optional, get and send pyAFQ rerun metrics instead of all.
+        rescan: Optional, get and send scan-rescan metrics instead of all.
 
     Raises:
         EnvironmentError: Method not executed on HCC.
@@ -509,14 +554,19 @@ def insert_pyafq(rerun: bool = False) -> pd.DataFrame:
             "workflows.setub_db is written for execution on HCC."
         )
 
-    deriv_dir = (
-        "/mnt/nrdstor/muncylab/nmuncy2/ADR/data_mri/derivatives"
-        if not rerun
-        else "/mnt/nrdstor/muncylab/nmuncy2/ADR/data_mri/derivatives_rerun"
-    )
+    # Identify relevant derivatives path
+    deriv_dir = "/mnt/nrdstor/muncylab/nmuncy2/ADR/data_mri/derivatives"
+    if rerun:
+        deriv_dir = (
+            "/mnt/nrdstor/muncylab/nmuncy2/ADR/data_mri/derivatives_rerun"
+        )
+    if rescan:
+        deriv_dir = "/mnt/nrdstor/muncylab/nmuncy2/scan_rescan/derivatives"
+
+    # Find data, insert it
     csv_path = os.path.join(deriv_dir, "afq", "tract_profiles.csv")
     if not os.path.exists(csv_path):
         raise FileNotFoundError(csv_path)
 
     df = pd.read_csv(csv_path)
-    return database.build_afq(df, rerun=rerun)
+    return database.build_afq(df, rerun=rerun, rescan=rescan)
